Move pause/suspend/resume functions to `flow_runs` module
